fix broken file structure and move all npm modules into root along with package.json and index.js
